Created RewardValue Class

+++ src/test/java/RewardValueTests.java
@@ -1,19 +1,20 @@
 import org.junit.jupiter.api.Test;
 
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class RewardValueTests {
 
     @Test
     void create_with_cash_value() {
-        double cashValue = 100;
+        double cashValue = 100.0;
         var rewardValue = new RewardValue(cashValue);
         assertEquals(cashValue, rewardValue.getCashValue());
     }
 
     @Test
     void create_with_miles_value() {
-        int milesValue = 10000;
+        double milesValue = 10000;
         var rewardValue = new RewardValue(milesValue);
         assertEquals(milesValue, rewardValue.getMilesValue());
     }
